Kill BasicStore, FullStore is now the only Store
